+++ integration_tests/suite/helpers/helpers/agent.py
@@ -1,10 +1,10 @@
 # -*- coding: utf-8 -*-
-# Copyright 2016-2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2016-2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
 import random
 
-from . import db
+from . import confd
 
 
 def generate_agent(**parameters):
@@ -13,26 +13,24 @@ def generate_agent(**parameters):
 
 
 def add_agent(**parameters):
-    with db.queries() as queries:
-        id = queries.insert_agent(**parameters)
-    return {'id': id,
-            'number': parameters.get('number')}
+    response = confd.agents.post(parameters)
+    return response.item
 
 
 def delete_agent(agent_id, check=False):
-    with db.queries() as queries:
-        queries.delete_agent(agent_id)
+    response = confd.agents(agent_id).delete()
+    if check:
+        response.assert_ok()
 
 
 def generate_number():
-    with db.queries() as queries:
-        response = queries.get_agent()
-    numbers = set(d['number'] for d in response)
+    response = confd.agents.get()
+    numbers = set(d['number'] for d in response.items)
     return _random_number(numbers)
 
 
 def _random_number(numbers):
     number = ''.join(random.choice('0123456789') for i in range(4))
-    while number in numbers:
-        number = ''.join(random.choice('0123456789') for i in range(4))
+    if number in numbers:
+        return _random_number(numbers)
     return number