+++ xivo_confd/plugins/agent/plugin.py
@@ -1,0 +1,26 @@
+# -*- coding: utf-8 -*-
+# Copyright 2018 The Wazo Authors  (see the AUTHORS file)
+# SPDX-License-Identifier: GPL-3.0+
+
+from .resource import AgentItem, AgentList
+from .service import build_service
+
+
+class Plugin(object):
+
+    def load(self, dependencies):
+        api = dependencies['api']
+        service = build_service()
+
+        api.add_resource(
+            AgentList,
+            '/agents',
+            resource_class_args=(service,)
+        )
+
+        api.add_resource(
+            AgentItem,
+            '/agents/<int:id>',
+            endpoint='agents',
+            resource_class_args=(service,)
+        )